feat: DimLoader 구현 [TRL-148]
